Add tiles.hipsdescription module
This commits adds a class named HipsDescription which provides methods for parsing the HiPS properties.
It parses the given string and returns an object of OrderedDict. Additional getter methods are also included.
For the test cases, a sample properties file was obtained from http://alasky.unistra.fr/DSS/DSSColor/properties